+++ modules/Grades/Transcripts.php
@@ -56,7 +56,7 @@ if ( $_REQUEST['modfunc'] === 'save' )
 			BackPrompt(
 				sprintf(
 					_( 'No %s were found.' ),
-					_( 'Grades' )
+					ngettext( 'Grade', 'Grades', 0 )
 				)
 			);
 		}